fix: make sure we don't miss dimensions updates
